New project structure
This new structure introduces support for csound 7, now that the API seems
to be taking shape. The differences between csound 6 and 7 are big enough
that full support for both via if clauses seems difficult. The csound 6
and 7 apis are now fully separated in two different modules. Great
effort is taken to make this update backwards compatible, by keeping
the naming convention of the csound 6 version (cammelCase) and introducing
compatibility methods to make possible to support both versions simultaneously.

With this new version we drift away from ctcsound. This means more support
burden in the future, but better control and hopefully better decisions.